fix capactiro polarized according to IPC
