+++ symbols/capacitor_polarized_down.ts
@@ -9,26 +9,26 @@ export default modifySymbol({
     {
       type: "text",
       text: "{REF}",
-      x: -0.2,
-      y: 0.2094553499999995,
+      x: 0.1,
+      y: -0.2094553499999995,
     },
     {
       type: "text",
       text: "{VAL}",
-      x: 0.3,
-      y: 0.2094553499999995,
+      x: -0.2,
+      y: -0.2094553499999995,
     },
   ] as any,
   ports: [
-    { ...refblocks.left1, labels: ["1"] }, // TODO add more "standard" labels
-    { ...refblocks.right1, labels: ["2"] }, // TODO add more "standard" labels
+    { ...refblocks.left1, labels: ["2"] }, // TODO add more "standard" labels
+    { ...refblocks.right1, labels: ["1"] }, // TODO add more "standard" labels
   ],
   size: { width: bounds.width, height: bounds.height },
   center: { x: bounds.centerX, y: bounds.centerY },
 })
   .changeTextAnchor("{VAL}", "top_left")
-  .rotateRightFacingSymbol("down")
-  .labelPort("left1", ["1", "neg"])
-  .labelPort("right1", ["2", "pos"])
+  .rotateRightFacingSymbol("up")
+  .labelPort("left1", ["1", "pos"])
+  .labelPort("right1", ["2", "neg"])
   .changeTextAnchor("{REF}", "bottom_left")
   .build()
