+++ symbols/capacitor_polarized_right.ts
@@ -1,10 +1,34 @@
-import { modifySymbol } from "../drawing/modify-symbol/modify-symbol"
 import svgJson from "assets/generated/capacitor_polarized.json"
+import { modifySymbol } from "drawing/modify-symbol/modify-symbol"
 
-export default modifySymbol(svgJson)
-  .changeTextAnchor("{VAL}", "middle_top")
-  .rotateRightFacingSymbol("right")
-  .labelPort("left1", ["1", "neg"])
-  .labelPort("right1", ["2", "pos"])
+const { paths, texts, bounds, refblocks, circles } = svgJson
+export default modifySymbol({
+  primitives: [
+    ...Object.values(paths),
+    ...Object.values(circles),
+    {
+      type: "text",
+      text: "{REF}",
+      x: 0,
+      y: -0.4094553499999995,
+    },
+    {
+      type: "text",
+      text: "{VAL}",
+      x: -0,
+      y: 0.4094553499999995,
+    },
+  ] as any,
+  ports: [
+    { ...refblocks.left1, labels: ["2"] }, // TODO add more "standard" labels
+    { ...refblocks.right1, labels: ["1"] }, // TODO add more "standard" labels
+  ],
+  size: { width: bounds.width, height: bounds.height },
+  center: { x: bounds.centerX, y: bounds.centerY },
+})
+  .changeTextAnchor("{VAL}", "middle_bottom")
+  .rotateRightFacingSymbol("left")
+  .labelPort("left1", ["2", "neg"])
+  .labelPort("right1", ["1", "pos"])
   .changeTextAnchor("{REF}", "middle_bottom")
   .build()
